Added UML - spent some time thinking about how to properly deal with object oriented design
This UML is not describing the code exactly, but I think I want to make most of the "actions" handled by the action class. In other words the action class gets passed the brain of a player, and then returns whatever it needs to.
Changes in code needed. Plans to iteratively change code then update UML.